requirements.txt
requests
beautifulsoup4
lxml
pandas
mysql-connector-python
